--- piikii-application/src/main/kotlin/com/piikii/application/domain/model/room/Room.kt
@@ -1,3 +1,0 @@
-package com.piikii.application.domain.model.room
-
-class Room
